created kubeconfig package
